beta3

+++ src/main/java/personal/pan/dl4j/nn/conf/graph/AttentionVertex.java
@@ -35,7 +35,7 @@ public class AttentionVertex extends GraphVertex {
 	}
 
 	@Override
-	public int numParams(boolean backprop) {
+	public long numParams(boolean backprop) {
 		return 0;
 	}
